Add MemoryParatextProjectTextUpdater and move MemoryParatextProjectFileHandler to production code
- Create MemoryParatextProjectFileHandler in machine/corpora/ (promoted from testutils)
- Create MemoryParatextProjectTextUpdater in machine/corpora/
- Rename testutils file to default_paratext_project_settings.py
- Update all imports to use new locations
- Refactor test_update_usfm_parser_handler to use MemoryParatextProjectTextUpdater
