+++ tests/testutils/memory_paratext_project_terms_parser.py
@@ -1,8 +1,8 @@
 from typing import Dict, Optional
 
-from machine.corpora import ParatextProjectSettings, ParatextProjectTermsParserBase
+from machine.corpora import MemoryParatextProjectFileHandler, ParatextProjectSettings, ParatextProjectTermsParserBase
 
-from .memory_paratext_project_file_handler import DefaultParatextProjectSettings, MemoryParatextProjectFileHandler
+from .default_paratext_project_settings import DefaultParatextProjectSettings
 
 
 class MemoryParatextProjectTermsParser(ParatextProjectTermsParserBase):
