+++ tests/testutils/memory_paratext_project_versification_error_detector.py
@@ -1,8 +1,12 @@
 from typing import Dict, Optional
 
-from machine.corpora import ParatextProjectSettings, ParatextProjectVersificationErrorDetectorBase
+from machine.corpora import (
+    MemoryParatextProjectFileHandler,
+    ParatextProjectSettings,
+    ParatextProjectVersificationErrorDetectorBase,
+)
 
-from .memory_paratext_project_file_handler import DefaultParatextProjectSettings, MemoryParatextProjectFileHandler
+from .default_paratext_project_settings import DefaultParatextProjectSettings
 
 
 class MemoryParatextProjectVersificationErrorDetector(ParatextProjectVersificationErrorDetectorBase):
